# net/textproto: eliminate some bounds checks